Splitting Planning
